# openspec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PPSC MCQ Power Bank",
    "version": "0.0.1"
  },
  "paths": {
...
    "/papers/": {
      "get": {
        "tags": [
          "Papers"
        ],
        "summary": "List Papers",
        "operationId": "list_papers_papers__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PapersListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Papers"
        ],
        "summary": "Generate a new paper",
        "description": "Creates a new paper by selecting the next `size` distinct questions that have appeared at least `threshold` times, ensuring no overlap with existing papers.",
        "operationId": "create_paper_papers__post",
        "parameters": [
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Minimum times a question must repeat",
              "default": 10,
              "title": "Threshold"
            },
            "description": "Minimum times a question must repeat"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Number of questions in the paper",
              "default": 100,
              "title": "Size"
            },
            "description": "Number of questions in the paper"
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Paper difficulty (default: medium)",
              "default": "medium",
              "title": "Difficulty"
            },
            "description": "Paper difficulty (default: medium)"
          },
          {
            "name": "category_slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category slug",
              "title": "Category Slug"
            },
            "description": "Filter by category slug"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category id",
              "title": "Category Id"
            },
            "description": "Filter by category id"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/papers/{paper_id}": {
      "get": {
        "tags": [
          "Papers"
        ],
        "summary": "Fetch one paper",
        "operationId": "get_paper_papers__paper_id__get",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/views/papers": {
      "get": {
        "tags": [
          "Papers View"
        ],
        "summary": "HTML view of papers list",
        "operationId": "papers_list_views_papers_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/views/papers/{paper_id}": {
      "get": {
        "tags": [
          "Papers View"
        ],
        "summary": "HTML view of one paper",
        "operationId": "paper_detail_views_papers__paper_id__get",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Paper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/paper-series": {
      "post": {
        "tags": [
          "Paper Series"
        ],
        "summary": "Create Series",
        "operationId": "create_series_paper_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperSeriesCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSeriesRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/paper-series/{series_id}": {
      "get": {
        "tags": [
          "Paper Series"
        ],
        "summary": "Get Series",
        "operationId": "get_series_paper_series__series_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSeriesRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paper-series/{series_id}/days": {
      "get": {
        "tags": [
          "Paper Series"
        ],
        "summary": "List Series Days",
        "operationId": "list_series_days_paper_series__series_id__days_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaperSeriesDayRead"
                  },
                  "title": "Response List Series Days Paper Series  Series Id  Days Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paper-series/{series_id}/today": {
      "get": {
        "tags": [
          "Paper Series"
        ],
        "summary": "Get Today Syllabus",
        "operationId": "get_today_syllabus_paper_series__series_id__today_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Id"
            }
          },
          {
            "name": "for_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "For Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSeriesDayRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paper-series/day/{day_id}/start": {
      "post": {
        "tags": [
          "Paper Series"
        ],
        "summary": "Start Day Attempt",
        "operationId": "start_day_attempt_paper_series_day__day_id__start_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "day_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Day Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSeriesAttemptStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PapersListResponse": {
        "properties": {
          "total_papers": {
            "type": "integer",
            "title": "Total Papers"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          },
          "papers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PaperItem"
            },
            "type": "object",
            "title": "Papers"
          }
        },
        "type": "object",
        "required": [
          "total_papers",
          "total_pages",
          "page",
          "per_page",
          "papers"
        ],
        "title": "PapersListResponse"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://ppsc-paper-bank.onrender.com",
      "description": "Render Production"
    }
  ]
}